refactor: 오해를 불러 일으킬 부적절한 이름을 적절하게 수정
