+++ src/main/kotlin/domain/store/LottoNumberGenerator.kt
@@ -1,0 +1,7 @@
+package domain.store
+
+import domain.lotto.LottoNumbers
+
+interface LottoNumberGenerator {
+    fun generate(): LottoNumbers
+}
