+++ src/main/kotlin/domain/store/RandomLottoNumberGenerator.kt
@@ -1,7 +1,12 @@
 package domain.store
 
+import domain.lotto.LottoNumber
 import domain.lotto.LottoNumbers
 
-interface RandomLottoNumberGenerator {
-    fun generate(): LottoNumbers
+class RandomLottoNumberGenerator : LottoNumberGenerator {
+    override fun generate(): LottoNumbers {
+        return LottoNumber.VALUES.shuffled()
+            .take(LottoNumbers.SIZE)
+            .let { LottoNumbers.fromList(it) }
+    }
 }
