+++ src/test/kotlin/domain/store/RandomLottoNumberGeneratorTest.kt
@@ -5,20 +5,20 @@ import org.assertj.core.api.Assertions.assertThat
 import org.junit.jupiter.api.Test
 import org.junit.jupiter.api.assertDoesNotThrow
 
-internal class CommonRandomLottoNumberGeneratorTest {
+internal class RandomLottoNumberGeneratorTest {
     @Test
     fun `무작위로또숫자생성기는 생성 시 아무 에러를 발생하지 않는다`() {
-        assertDoesNotThrow { CommonRandomLottoNumberGenerator() }
+        assertDoesNotThrow { RandomLottoNumberGenerator() }
     }
 
     @Test
     fun `무작위로또숫자생성기는 로또숫자열을 하나 생성한다`() {
-        assertThat(CommonRandomLottoNumberGenerator().generate()).isInstanceOf(LottoNumbers::class.java)
+        assertThat(RandomLottoNumberGenerator().generate()).isInstanceOf(LottoNumbers::class.java)
     }
 
     @Test
     fun `무작위로또숫자생성기는 로또숫자열의 최대 개수만큼 생성한다`() {
-        val quickPick = CommonRandomLottoNumberGenerator().generate()
+        val quickPick = RandomLottoNumberGenerator().generate()
         assertThat(quickPick.toList()).hasSize(LottoNumbers.SIZE)
     }
 }
